+++ ChocolateyPackageCreator/functions/publish.ps1
@@ -11,8 +11,12 @@
     A NuGet API key with sufficient privileges to publish the package
 .PARAMETER PackageFile
     The path to the Chocolatey package file
+.PARAMETER Tool
+    The tool to use to perform the publish: Chocolatey or NuGet
 .PARAMETER ChocolateyPath
     The path to the choco binary - defaults to 'choco'
+.PARAMETER NuGetPath
+    The path to the NuGet binary - defaults to 'nuget'
 .PARAMETER ExtraArguments
     Additional arguments to pass to the choco binary during the push process
 .PARAMETER Force
@@ -44,31 +48,48 @@ Function Publish-ChocolateyPackage {
             Position = 3
         )]
         [string] $PackageFile,
+        [ValidateSet('Chocolatey', 'NuGet')]
+        [string] $Tool = 'Chocolatey',
         [string] $ChocolateyPath = 'choco',
+        [string] $NuGetPath = 'nuget',
         [string[]] $ExtraArguments = @(),
         [switch] $Force
     )
 
     Write-Verbose ('Pushing Chocolatey package at {0}...' -f $PackageFile)
-    $chocoArgs = [System.Collections.ArrayList]@(
-        'push',
-        $PackageFile,
-        ('--source "{0}"' -f $Repository),
-        ('--api-key "{0}"' -f $ApiKey)
-    )
+    switch ($Tool) {
+        Chocolatey {
+            $toolPath = $ChocolateyPath
+            $toolArgs = [System.Collections.ArrayList]@(
+                'push',
+                $PackageFile,
+                ('--source "{0}"' -f $Repository),
+                ('--api-key "{0}"' -f $ApiKey)
+            )
 
-    if ($Force) {
-        $chocoArgs.Add('--force') | Out-Null
+            if ($Force) {
+                $toolArgs.Add('--force') | Out-Null
+            }
+        }
+        NuGet {
+            $toolPath = $NuGetPath
+            $toolArgs = [System.Collections.ArrayList]@(
+                'push',
+                $PackageFile,
+                ('-source "{0}"' -f $Repository),
+                ('-apikey "{0}"' -f $ApiKey)
+            )
+        }
     }
 
     if ($ExtraArguments) {
-        $chocoArgs += $ExtraArguments
+        $toolArgs += $ExtraArguments
     }
-    
+
     # Don't print out sensitive information
-    $logStr = "Executing `"{0} {1}`"" -f $ChocolateyPath, ($chocoArgs -join ' ')
+    $logStr = "Executing `"{0} {1}`"" -f $toolPath, ($toolArgs -join ' ')
     Write-Verbose ($logStr -replace $ApiKey, '<REDACTED>')
 
-    $proc = Start-Process $ChocolateyPath -ArgumentList $chocoArgs -PassThru -NoNewWindow -Wait
+    $proc = Start-Process $toolPath -ArgumentList $toolArgs -PassThru -NoNewWindow -Wait
     $proc.ExitCode
 }